requirements.txt
beautifulsoup4
requests
tqdm
sentence-transformers
faiss-cpu

